mcp-server/pyproject.toml
[project]
name = "pyrsm-mcp-server"
version = "0.1.0"
description = "MCP server for pyrsm - exposes pyrsm tools to AI assistants"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp>=1.0.0",
    "nest-asyncio>=1.6.0",
    "pyrsm>=1.6.0",
]

[project.scripts]
pyrsm-mcp = "server:main"
